requirements.txt
cm-api
python-ldap
